+++ ui/great-building.js
@@ -48,7 +48,7 @@ function updateGreatBuildingChanges (changes) {
     ggRows = '<tr><td colspan="3">No buildings</td></tr>'
   } else {
     if (changes.length === 1 && changes[0].name === 'last change') {
-      ggRows = '<tr><td colspan="3">Last change ' + moment.unix(changes[0].last_spent).fromNow() + '</td></tr>'
+      ggRows = '<tr><td colspan="3">' + (changes[0].last_spent ? 'Last change ' + moment.unix(changes[0].last_spent).fromNow() : 'Not available') + '</td></tr>'
     } else {
       for (let i = 0; i < changes.length; i++) {
         ggRows += addGreatBuildingChangesRow(changes[i])
@@ -125,5 +125,5 @@ function addGreatBuildingBoostRow (spot, boostFactor, analysis, totalFP, freeFP,
 }
 
 function addGreatBuildingChangesRow (change) {
-  return '<tr><td>' + change.name + '</td><td>' + moment.unix(change.last_spent).fromNow(true) + '</td><td>' + change.completePercentage + '%</td></tr>'
+  return '<tr><td>' + change.name + '</td><td>' + (change.last_spent ? moment.unix(change.last_spent).fromNow(true) : '<em>closed</em>') + '</td><td>' + change.completePercentage + '%</td></tr>'
 }