+++ CHANGELOG.md
@@ -1,4 +1,18 @@
 # Upcoming
+* The Crow's Nest cannot be plundered
+* Show 'not available' or 'closed' if the last change info of a GB is not available ([#10](https://github.com/veger/foei/issues/10))
+* Show if battle was 'automatic'
+
+# 0.4.0
+* Add GvG tab (showing sector army details)
+* Fix GB calculation when self (still unsafe) is followed by owner
+* Fix (auto) battle info for newest version of FoE
+* Show images of buildings in plunder tab (thanks for contribution)
+* More building and unit IDs
+
+# 0.3.2
+* Fix battle info for newest version of FoE
+* More building and unit IDs
 
 # 0.3.1
 * Fix GB calculation when self is followed by owner